requirements.txt
asgiref
astroid
Django
djangorestframework
django-filter
isort
lazy-object-proxy
mccabe
pylint
pytz
six
sqlparse
toml
typed-ast
wrapt
validate-docbr